+++ src/decorators/deferHandler.js
@@ -1,10 +1,12 @@
 import { apply, curry, equals, ifElse, tap } from 'ramda';
 import { isFunction } from 'ramda-adjunct';
+import { getHandler, getId, setHandler } from '../lens';
 import pipeAsync from '../util/pipeAsync';
-import { debugItRuntime } from '../util/debug';
+import { createDebug } from '../util/debug';
 import invariant from '../util/invariant';
-import { getHandler, getId, setHandler } from '../lens/feature';
 import { HANDLER, HANDLER_NOT_READY_RESULT } from '../constants';
+
+const debugIt = createDebug('deferHandler');
 
 const wrapHandler = curry((validator, whenPassed, whenFailed) => (...args) =>
   pipeAsync(
@@ -26,7 +28,7 @@ const deferHandler = curry((validator, feature) => {
     HANDLER,
   );
 
-  const whenFailed = () => tap(() => debugItRuntime('defered', getId(feature)));
+  const whenFailed = () => tap(() => debugIt('defered', getId(feature)));
   const whenPassed = args => () => originalHandler(...args);
   const newHandler = wrapHandler(validator, whenPassed, whenFailed);
