--- src/util/promise.js
@@ -1,6 +1,0 @@
-import { andThen, pipe, unless } from 'ramda';
-import { allP, isPromise, resolveP } from 'ramda-adjunct';
-
-export const ensurePromise = unless(isPromise, resolveP);
-
-export const promiseAllAndThen = fn => pipe(allP, andThen(fn));
